requirements.txt
flask>=2.3
opencv-python-headless>=4.8
Pillow>=10.0
imagehash>=4.3
requests>=2.31
werkzeug>=2.3
numpy>=1.24
google-generativeai>=0.8